+++ test/FESpacesTests/CLagrangianFESpacesTests.jl
@@ -73,11 +73,11 @@ V = CLagrangianFESpace(VectorValue{2,Float64},grid,Vector{Float64},node_to_tag,m
 
 reffe = ReferenceFE(lagrangian,Float64,2)
 V = FESpace(model,reffe)
-@test V.metadata === nothing
+@test isa(V.metadata, FESpaces.CellConformity)
 
 reffe = ReferenceFE(lagrangian,Float64,1)
 V = FESpace(model,reffe,conformity=:L2)
-@test V.metadata === nothing
+@test isa(V.metadata, FESpaces.CellConformity)
 
 # Check that the factory uses clagrangian when possible
 
@@ -107,6 +107,6 @@ partition = (3,3)
 model = CartesianDiscreteModel(domain,partition,isperiodic=(false,true))
 reffe = ReferenceFE(lagrangian,Float64,1)
 V = FESpace(model,reffe)
-@test V.metadata === nothing
+@test isa(V.metadata, FESpaces.CellConformity)
 
 end # module